+++ kmock-processor/src/test/resources/mock/template/generic/SuperTyped.kt
@@ -27,12 +27,16 @@ interface Parent<K, L> where L : Any, L : Comparable<L>, K : Any {
     fun <T, K> lol(x: T) where T : K
 
     fun narv(vararg x: L)
+    fun narv(vararg x: Array<out Any>)
+    fun narv(vararg x: IntArray)
 
     fun oo(fuzz: Int, vararg ozz: Any): Any
     fun oo(fuzz: Any?, vararg ozz: Int): Any
     fun <T> oo(fuzz: Any, vararg ozz: T): Any
-    fun <T> oo(fuzz: Any, vararg ozz: RRR<T>): Any
+    fun <T> oo(fuzz: Any, vararg ozz: RRR<out T>): Any
     fun <T> oo(fuzz: Any, ozz: CharArray): Any
+
+    fun <T> roo(fuzz: Any, vararg ozz: RRR<*>): Any
 }
 
 class RRR<L>